Replace "new chat" screen with a modal
