+++ src/screens/Chat/ChatList.tsx
@@ -5,7 +5,7 @@ import {
   useFonts,
 } from "@expo-google-fonts/poppins";
 import { StackScreenProps } from "@react-navigation/stack";
-import React from "react";
+import React, { useState } from "react";
 import { FlatList, StyleSheet, View } from "react-native";
 
 import { Params } from ".";
@@ -13,10 +13,12 @@ import { chats } from "../../_data/chats";
 import { ChatListItem } from "../../components/Chat/ChatListItem";
 import { NewChatButton } from "../../components/NewChatButton";
 import { GradientHeader } from "../../components/GradientHeader";
+import { ChatNewModal } from "./ChatNewModal";
 
 type Props = StackScreenProps<Params, "ChatList">;
 
 const ChatList: React.FC<Props> = ({ navigation }) => {
+  const [newChatVisible, setNewChatVisible] = useState(false);
   const [fontsLoaded] = useFonts({
     Montserrat_600SemiBold,
     Poppins_400Regular,
@@ -44,7 +46,12 @@ const ChatList: React.FC<Props> = ({ navigation }) => {
           />
         )}
       />
-      <NewChatButton onPress={() => navigation.navigate("ChatNew", {})} />
+      <NewChatButton onPress={() => setNewChatVisible(true)} />
+      <ChatNewModal
+        visible={newChatVisible}
+        onClose={() => setNewChatVisible(false)}
+        navigateToChat={(id) => navigation.navigate("ChatView", { id })}
+      />
     </View>
   );
 };
